Feature: Store and provide more product details

+++ src/main/java/br/com/sales/micro/dto/request/ProductBasicInfoDto.java
@@ -2,6 +2,6 @@ package br.com.sales.micro.dto.request;
 
 public record ProductBasicInfoDto(
         Long productBarCode,
-        Long productQuantity
+        Integer productQuantity
 ) {
 }
